+++ docs/install/README.md
@@ -339,6 +339,10 @@ gcloud app deploy
 
     -   `GITHUB_ACCESS_TOKEN`: GitHub Access Token
 
+-   `bilibili`:
+
+    -   `BILIBILI_COOKIE_{uid}`: id 为{uid}的 b 站用户登录后的 Cookie 值
+
 ### 访问控制
 
 可以通过修改 `middleware/access-control.js` 或者设置环境变量来配置黑名单和白名单.
