+++ lib/config.js
@@ -49,6 +49,10 @@ module.exports = {
         name: process.env.HTTP_BASIC_AUTH_NAME || 'usernam3',
         pass: process.env.HTTP_BASIC_AUTH_PASS || 'passw0rd',
     },
+    bilibili: {
+        cookies: {},
+    },
+
     puppeteerWSEndpoint: process.env.PUPPETEER_WS_ENDPOINT,
     loggerLevel: process.env.LOGGER_LEVEL || 'info',
     proxy: {
@@ -61,3 +65,11 @@ module.exports = {
     blacklist: process.env.BLACKLIST && process.env.BLACKLIST.split(','),
     whitelist: process.env.WHITELIST && process.env.WHITELIST.split(','),
 };
+
+const envs = process.env;
+for (const name in envs) {
+    if (name.startsWith('BILIBILI_COOKIE_')) {
+        const uid = name.slice(16);
+        module.exports.bilibili.cookies[uid] = envs[name];
+    }
+}
