+++ lib/router.js
@@ -120,7 +120,7 @@ router.get('/bilibili/user/channel/:uid/:cid', require('./routes/bilibili/userCh
 router.get('/bilibili/topic/:topic', require('./routes/bilibili/topic'));
 router.get('/bilibili/audio/:id', require('./routes/bilibili/audio'));
 router.get('/bilibili/vsearch/:kw/:order?', require('./routes/bilibili/vsearch'));
-router.get('/bilibili/followings/video/:uid', require('./routes/bilibili/followings_video'));
+router.get('/bilibili/dynamic/video/:uid', require('./routes/bilibili/dynamic_video'));
 
 // bangumi
 router.get('/bangumi/calendar/today', require('./routes/bangumi/calendar/today'));
