+++ lib/routes/bilibili/dynamic_video.js
@@ -1,0 +1,40 @@
+const axios = require('../../utils/axios');
+const cache = require('./cache');
+const config = require('../../config');
+
+module.exports = async (ctx) => {
+    const uid = String(ctx.params.uid);
+    const name = await cache.getUsernameFromUID(ctx, uid);
+
+    const cookie = config.bilibili.cookies[uid];
+
+    const response = await axios({
+        method: 'get',
+        url: `https://api.vc.bilibili.com/dynamic_svr/v1/dynamic_svr/dynamic_new?uid=${uid}&type=8`,
+        headers: {
+            Referer: `https://space.bilibili.com/${uid}/`,
+            Cookie: cookie,
+        },
+    });
+    const cards = response.data.data.cards;
+
+    const out = await Promise.all(
+        cards.map(async (card) => {
+            const card_data = JSON.parse(card.card);
+
+            const item = {
+                title: card_data.title,
+                description: `${card_data.desc}<br><img referrerpolicy="no-referrer" src="${card_data.pic}">`,
+                pubDate: new Date(card_data.pubdate * 1000).toUTCString(),
+                link: `https://www.bilibili.com/video/av${card_data.aid}`,
+            };
+            return Promise.resolve(item);
+        })
+    );
+
+    ctx.state.data = {
+        title: `${name} 动态首页投稿视频`,
+        link: `https://t.bilibili.com/?tab=8`,
+        item: out,
+    };
+};
